+++ Application/LogComponent/ILogger.cs
@@ -1,3 +1,5 @@
+using System.Threading.Tasks;
+
 namespace Application.LogComponent
 {
 	public interface ILogger
@@ -5,12 +7,12 @@ namespace Application.LogComponent
 		/// <summary>
 		/// Stop the logging. If there are any outstanding logs, these will not be written to the Log.
 		/// </summary>
-		void Stop_Without_Flush();
+		void StopWithoutFlush();
 
 		/// <summary>
 		/// Stop the logging. The call will not return until all logs have been written to the Log.
 		/// </summary>
-		void Stop_With_Flush();
+		Task StopWithFlush();
 
 		/// <summary>
 		/// Write a message to the Log.