+++ Application/Program.cs
@@ -1,24 +1,25 @@
 ﻿using System;
 using System.Threading;
+using System.Threading.Tasks;
 using Application.LogComponent;
 
 namespace Application
 {
 	class Program
 	{
-		static void Main(string[] args)
+		static async Task Main(string[] args)
 		{
 			TimeProvider timeProvider = TimeProvider.System;
 
-			ILogger logger = new AsyncLogger(timeProvider);
+			ILogger logger1 = new AsyncLogger(timeProvider);
 
 			for (int i = 0; i < 15; i++)
 			{
-				logger.WriteLog("Number with flush: " + i.ToString());
+				logger1.WriteLog("Number with flush: " + i.ToString());
 				Thread.Sleep(50);
 			}
 
-			logger.Stop_With_Flush();
+			var logger1Thread = Task.Run(() => logger1.StopWithFlush());
 
 			ILogger logger2 = new AsyncLogger(timeProvider);
 
@@ -28,7 +29,9 @@ namespace Application
 				Thread.Sleep(20);
 			}
 
-			logger2.Stop_Without_Flush();
+			logger2.StopWithoutFlush();
+
+			await logger1Thread;
 
 			//Console.ReadLine();
 		}